requirements.txt
# pip install torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu117
pandas
matplotlib
scikit-learn
numpy==1.25.0
tqdm==4.65.0
transformers==4.33.2
lightning==2.0.9
pillow==9.5.0
torchsummary
ssqueezepy
tensorboard
